--- lua/bootstrap.lua
@@ -1,7 +1,0 @@
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-if not vim.loop.fs_stat(lazypath) then
-  print("installing lazy.nvim..")
-  vim.fn.system({ "git", "clone", "--filter=blob:none", "https://github.com/folke/lazy.nvim.git", lazypath })
-  vim.fn.system({ "git", "-C", lazypath, "checkout", "tags/stable" }) -- last stable release
-end
-vim.opt.rtp:prepend(lazypath)
