+++ lua/e/plugins/comment.lua
@@ -1,0 +1,9 @@
+return {
+   {
+     "numToStr/Comment.nvim",
+    lazy = false,
+     config = function()
+       require("Comment").setup()
+     end,
+   }
+}
