+++ lua/e/plugins/core.lua
@@ -1,0 +1,4 @@
+return {
+  "nvim-lua/plenary.nvim",
+  "nvim-tree/nvim-web-devicons",
+}
