+++ lua/e/plugins/lsp/servers.lua
@@ -1,0 +1,23 @@
+return {
+  sumneko_lua = {
+    settings = {
+      Lua = {
+        workspace = {
+          checkThirdParty = false,
+        },
+        completion = {
+          callSnippet = "Replace",
+        },
+      },
+    },
+  },
+  pyright = {},
+  tsserver = {},
+  bashls = {},
+  yamlls = {},
+  jsonls = {},
+  sqls = {},
+  terraformls = {},
+  gopls = {},
+  dockerls = {},
+}
