+++ lua/e/plugins/misc.lua
@@ -1,0 +1,3 @@
+return {
+  { "j-hui/fidget.nvim", config = true },
+}
