--- lua/e/telescope.lua
@@ -1,19 +1,0 @@
--- search module
-local telescope = require("telescope")
-telescope.setup({
-  defaults = {
-    vimgrep_arguments = {
-      "rg",
-      "-L",
-      "--color=never",
-      "--no-heading",
-      "--with-filename",
-      "--line-number",
-      "--column",
-      "--smart-case",
-    },
-  },
-})
-
--- extensions
-require("telescope").load_extension("ui-select")
